Ajuste do tipo do campos step.

+++ zenon-fraud-detector/src/main/java/br/com/zenon/fraud/models/Transaction.java
@@ -2,7 +2,7 @@ package br.com.zenon.fraud.models;
 
 import java.math.BigDecimal;
 
-public record Transaction(long step, Type type, BigDecimal amount,
+public record Transaction(int step, Type type, BigDecimal amount,
                           String nameOrig, BigDecimal oldbalanceOrg, BigDecimal newbalanceOrig,
                           String nameDest, BigDecimal oldbalanceDest, BigDecimal newbalanceDest,
                           byte isFraud, byte isFlaggedFraud) {
